refactor: extract reusable formatting
